requirements.txt
pdfminer.six==20241021
